# Speed up test_pack.py by `fuse()`ing once per test case.